+++ src/pygpsclient/serverconfig_dialog.py
@@ -1,7 +1,7 @@
 """
-serverconfig_frame.py
+serverconfig_dialog.py
 
-Socket Server / NTRIP caster configuration panel Frame class.
+Socket Server / NTRIP caster configuration panel Dialog class.
 Supports two modes of operation - Socket Server and NTRIP Caster.
 
 If running in NTRIP Caster mode, two base station modes are available -
@@ -26,6 +26,7 @@ from tkinter import (
     DISABLED,
     EW,
     NORMAL,
+    NSEW,
     Button,
     Checkbutton,
     DoubleVar,
@@ -92,6 +93,7 @@ from pygpsclient.receiver_config_handler import (
 )
 from pygpsclient.strings import (
     DLGNOTLS,
+    DLGTSERVER,
     LBLACCURACY,
     LBLCONFIGBASE,
     LBLDISNMEA,
@@ -103,6 +105,7 @@ from pygpsclient.strings import (
     LBLSERVERPORT,
     LBLSOCKSERVE,
 )
+from pygpsclient.toplevel_dialog import ToplevelDialog
 
 ACCURACIES = (
     10.0,
@@ -134,34 +137,32 @@ POS_ECEF = "ECEF"
 POS_LLH = "LLH"
 PQTMVER = "PQTMVER"
 POSMODES = (POS_LLH, POS_ECEF)
+MINDIM = (400, 600)
 
 
-class ServerConfigFrame(Frame):
+class ServerConfigDialog(ToplevelDialog):
     """
-    Server configuration frame class.
+    Server configuration dialog class.
     """
 
-    def __init__(self, app, container, *args, **kwargs):
+    def __init__(self, app, *args, **kwargs):
         """
         Constructor.
 
         :param Frame app: reference to main tkinter application
-        :param Frame container: reference to container frame
         :param args: optional args to pass to Frame parent class
         :param kwargs: optional kwargs for value ranges, or to pass to Frame parent class
         """
 
-        Frame.__init__(self, container, *args, **kwargs)
-        self.logger = logging.getLogger(__name__)
-
         self.__app = app
-        self._container = container
+        self.logger = logging.getLogger(__name__)
+        super().__init__(app, DLGTSERVER, MINDIM)
+
         self._show_advanced = False
         self._socket_serve = IntVar()
         self.sock_port = StringVar()
         self.sock_host = StringVar()
         self.sock_mode = StringVar()
-        self._sock_clients = IntVar()
         self.receiver_type = StringVar()
         self.base_mode = StringVar()
         self.https = IntVar()
@@ -185,16 +186,18 @@ class ServerConfigFrame(Frame):
 
         self._body()
         self._do_layout()
-        self.reset()
+        self._reset()
         # self._attach_events() # done in reset
         self._attach_events1()
+        self._finalise()
 
     def _body(self):
         """
         Set up widgets.
         """
 
-        self._frm_basic = Frame(self)
+        self._frm_body = Frame(self.container, borderwidth=2, relief="groove")
+        self._frm_basic = Frame(self._frm_body)
         self._chk_socketserve = Checkbutton(
             self._frm_basic,
             text=LBLSOCKSERVE,
@@ -254,11 +257,6 @@ class ServerConfigFrame(Frame):
             relief="sunken",
             width=6,
         )
-        self._lbl_clients = Label(self._frm_basic, text="Clients")
-        self._lbl_sockclients = Label(
-            self._frm_basic,
-            textvariable=self._sock_clients,
-        )
         self._btn_toggle = Button(
             self._frm_basic,
             command=self._on_toggle_advanced,
@@ -267,7 +265,7 @@ class ServerConfigFrame(Frame):
             height=22,
             # state=DISABLED,
         )
-        self._frm_advanced = Frame(self)
+        self._frm_advanced = Frame(self._frm_body)
         self._lbl_user = Label(
             self._frm_advanced,
             text="User",
@@ -402,6 +400,7 @@ class ServerConfigFrame(Frame):
         Layout widgets.
         """
 
+        self._frm_body.grid(column=0, row=0, sticky=NSEW)
         self._frm_basic.grid(column=0, row=0, columnspan=5, sticky=EW)
         self._chk_socketserve.grid(
             column=0, row=0, columnspan=2, rowspan=2, padx=2, pady=1, sticky=W
@@ -410,8 +409,6 @@ class ServerConfigFrame(Frame):
         self._spn_sockmode.grid(column=3, row=0, padx=2, pady=1, sticky=W)
         self._lbl_sockhost.grid(column=0, row=2, padx=2, pady=1, sticky=W)
         self._ent_sockhost.grid(column=1, row=2, padx=2, pady=1, sticky=W)
-        self._lbl_clients.grid(column=2, row=2, padx=2, pady=1, sticky=W)
-        self._lbl_sockclients.grid(column=3, row=2, padx=2, pady=1, sticky=W)
         self._lbl_sockport.grid(column=0, row=3, padx=2, pady=1, sticky=W)
         self._ent_sockport.grid(column=1, row=3, padx=2, pady=1, sticky=W)
         self._chk_https.grid(column=2, row=3, columnspan=2, padx=2, pady=1, sticky=W)
@@ -427,14 +424,15 @@ class ServerConfigFrame(Frame):
         self._lbl_basemode.grid(column=0, row=1, padx=2, pady=1, sticky=E)
         self._spn_basemode.grid(column=1, row=1, padx=2, pady=1, sticky=W)
 
-    def reset(self):
+    def _reset(self):
         """
         Reset settings to defaults.
         """
 
         self._attach_events(False)
         cfg = self.__app.configuration
-        self._socket_serve.set(cfg.get("sockserver_b"))
+        # self._socket_serve.set(cfg.get("sockserver_b"))
+        self._socket_serve.set(self.__app.server_status >= 0)  # TODO
         self.sock_mode.set(SOCKMODES[cfg.get("sockmode_b")])
         self._on_toggle_advanced()
         self.base_mode.set(cfg.get("ntripcasterbasemode_s"))
@@ -453,7 +451,7 @@ class ServerConfigFrame(Frame):
         pem = cfg.get("tlspempath_s")
         if https and not Path(pem).exists():
             err = DLGNOTLS.format(hostpem=pem)
-            self.__app.status_label = (err, ERRCOL)
+            self.status_label = (err, ERRCOL)
             self.logger.error(err)
             cfg.set("sockhttps_b", 0)
             self._chk_https.config(state=DISABLED)
@@ -467,7 +465,6 @@ class ServerConfigFrame(Frame):
             self.sock_port.set(cfg.get("sockport_n"))
         self.user.set(cfg.get("ntripcasteruser_s"))
         self.password.set(cfg.get("ntripcasterpassword_s"))
-        self.clients = 0
         self._fixed_lat_temp = self.fixedlat.get()
         self._fixed_lon_temp = self.fixedlon.get()
         self._fixed_hae_temp = self.fixedhae.get()
@@ -559,7 +556,6 @@ class ServerConfigFrame(Frame):
         if status == DISCONNECTED:
             self._chk_socketserve.configure(state=DISABLED)
             self._socket_serve.set(0)
-            self.clients = 0
         else:
             self._chk_socketserve.configure(state=NORMAL)
 
@@ -569,9 +565,9 @@ class ServerConfigFrame(Frame):
         """
 
         if self.valid_settings():
-            self.__app.status_label = ("", INFOCOL)
+            self.status_label = ("", INFOCOL)
         else:
-            self.__app.status_label = ("ERROR - invalid entry", ERRCOL)
+            self.status_label = ("ERROR - invalid entry", ERRCOL)
             return
 
         self._quectel_restart = 0
@@ -585,7 +581,6 @@ class ServerConfigFrame(Frame):
         else:  # stop server
             self.__app.sockserver_stop()
             self.__app.stream_handler.sock_serve = False
-            self.clients = 0
 
         # set visibility of various fields depending on server status
         for wid in (
@@ -651,9 +646,9 @@ class ServerConfigFrame(Frame):
 
         # validate settings
         if self.valid_settings():
-            self.__app.status_label = ("", INFOCOL)
+            self.status_label = ("", INFOCOL)
         else:
-            self.__app.status_label = ("ERROR - invalid entry", ERRCOL)
+            self.status_label = ("ERROR - invalid entry", ERRCOL)
             return
 
         delay = self.__app.configuration.get("guiupdateinterval_f") / 2
@@ -864,7 +859,7 @@ class ServerConfigFrame(Frame):
         pem = self.__app.configuration.get("tlspempath_s")
         if self.https.get() and not Path(pem).exists():
             err = DLGNOTLS.format(hostpem=pem)
-            self.__app.status_label = (err, ERRCOL)
+            self.status_label = (err, ERRCOL)
             self.logger.error(err)
             self._attach_events(False)
             self.https.set(0)
@@ -1000,26 +995,6 @@ class ServerConfigFrame(Frame):
         self.fixedlat.set(lat)
         self.fixedlon.set(lon)
         self.fixedhae.set(hae)
-
-    @property
-    def clients(self) -> int:
-        """
-        Getter for number of socket clients.
-        """
-
-        return self._sock_clients.get()
-
-    @clients.setter
-    def clients(self, clients: int):
-        """
-        Setter for number of socket clients.
-
-        :param int clients: no of clients connected
-        """
-
-        self._sock_clients.set(clients)
-        if self._socket_serve.get() in ("1", 1):
-            self.__app.frm_banner.update_transmit_status(clients)
 
     def _config_msg_rates(self, rate: int, port_type: str):
         """