+++ src/pygpsclient/signalsview_frame.py
@@ -136,9 +136,6 @@ class SignalsviewFrame(Frame):
         self.__app.configuration.set(
             "unusedsat_b", not self.__app.configuration.get("unusedsat_b")
         )
-        self.__app.frm_settings.show_unusedsat.set(
-            self.__app.configuration.get("unusedsat_b")
-        )
         self._redraw = True
 
     def enable_messages(self, status: bool):